# 47% FastGen speedup for low workload - refactor allocator 